requirements.txt
fastapi
uvicorn
python-dotenv
openai
PyPDF2
